// src/api/api-client/isi-backend/models/BauratendateiInputDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ISI Backend API
 * ISI Backend - Service für das Informationssystem für soziale Infrastrukturplanung
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: noreply@mail.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues, parseDate, parseDateTime, serializeDate, serializeDateTime } from '../runtime';
import type { WohneinheitenProFoerderartProJahrDto } from './WohneinheitenProFoerderartProJahrDto';
import {
    WohneinheitenProFoerderartProJahrDtoFromJSON,
    WohneinheitenProFoerderartProJahrDtoFromJSONTyped,
    WohneinheitenProFoerderartProJahrDtoToJSON,
    WohneinheitenProFoerderartProJahrDtoToJSONTyped,
} from './WohneinheitenProFoerderartProJahrDto';

/**
 * 
 * @export
 * @interface BauratendateiInputDto
 */
export interface BauratendateiInputDto {
    /**
     * 
     */
    id?: string;
    /**
     * 
     */
    version?: number;
    /**
     * 
     */
    createdDateTime?: Date;
    /**
     * 
     */
    lastModifiedDateTime?: Date;
    /**
     * 
     */
    grundschulsprengel?: Array<string>;
    /**
     * 
     */
    mittelschulsprengel?: Array<string>;
    /**
     * 
     */
    viertel?: Array<string>;
    /**
     * 
     */
    wohneinheiten?: Array<WohneinheitenProFoerderartProJahrDto>;
}

/**
 * Check if a given object implements the BauratendateiInputDto interface.
 */
export function instanceOfBauratendateiInputDto(value: object): value is BauratendateiInputDto {
    return true;
}

export function BauratendateiInputDtoFromJSON(json: any): BauratendateiInputDto {
    return BauratendateiInputDtoFromJSONTyped(json, false);
}

export function BauratendateiInputDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BauratendateiInputDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'version': json['version'] == null ? undefined : json['version'],
        'createdDateTime': json['createdDateTime'] == null ? undefined : (parseDateTime(json['createdDateTime'])),
        'lastModifiedDateTime': json['lastModifiedDateTime'] == null ? undefined : (parseDateTime(json['lastModifiedDateTime'])),
        'grundschulsprengel': json['grundschulsprengel'] == null ? undefined : json['grundschulsprengel'],
        'mittelschulsprengel': json['mittelschulsprengel'] == null ? undefined : json['mittelschulsprengel'],
        'viertel': json['viertel'] == null ? undefined : json['viertel'],
        'wohneinheiten': json['wohneinheiten'] == null ? undefined : ((json['wohneinheiten'] as Array<any>).map(WohneinheitenProFoerderartProJahrDtoFromJSON)),
    };
}

export function BauratendateiInputDtoToJSON(json: any): BauratendateiInputDto {
    return BauratendateiInputDtoToJSONTyped(json, false);
}

export function BauratendateiInputDtoToJSONTyped(value?: BauratendateiInputDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'version': value['version'],
        'createdDateTime': value['createdDateTime'] == null ? value['createdDateTime'] : serializeDateTime(value['createdDateTime']),
        'lastModifiedDateTime': value['lastModifiedDateTime'] == null ? value['lastModifiedDateTime'] : serializeDateTime(value['lastModifiedDateTime']),
        'grundschulsprengel': value['grundschulsprengel'],
        'mittelschulsprengel': value['mittelschulsprengel'],
        'viertel': value['viertel'],
        'wohneinheiten': value['wohneinheiten'] == null ? undefined : ((value['wohneinheiten'] as Array<any>).map(WohneinheitenProFoerderartProJahrDtoToJSON)),
    };
}